requirements.txt
# Core dependencies
groq>=0.4.1
python-dotenv>=1.0.0
numpy>=1.24.0

# Optional but recommended
matplotlib>=3.7.0  # For visualization utilities

